Move classes to the python part of the project
Refs #29987
