+++ Framework/Kernel/CMakeLists.txt
@@ -104,8 +104,6 @@ set(SRC_FILES
     src/StartsWithValidator.cpp
     src/Statistics.cpp
     src/StdoutChannel.cpp
-    src/PythonStdoutChannel.cpp
-    src/PythonSink.cpp
     src/StringContainsValidator.cpp
     src/StringTokenizer.cpp
     src/Strings.cpp
@@ -280,7 +278,6 @@ set(INC_FILES
     inc/MantidKernel/StartsWithValidator.h
     inc/MantidKernel/Statistics.h
     inc/MantidKernel/StdoutChannel.h
-    inc/MantidKernel/PythonSink.h
     inc/MantidKernel/StringContainsValidator.h
     inc/MantidKernel/StringTokenizer.h
     inc/MantidKernel/Strings.h
@@ -429,7 +426,6 @@ set(TEST_FILES
     StartsWithValidatorTest.h
     StatisticsTest.h
     StdoutChannelTest.h
-    PythonStdoutChannelTest.h
     StringContainsValidatorTest.h
     StringTokenizerTest.h
     StringsTest.h
@@ -519,7 +515,6 @@ target_link_libraries(Kernel
                       ${JSONCPP_LIBRARIES}
                       ${TBB_LIBRARIES}
                       ${TBB_MALLOC_LIBRARIES}
-		      Python::Python
                       PRIVATE
                       ${GSL_LIBRARIES}
                       ${MANTIDLIBS}
