+++ Framework/PythonInterface/core/CMakeLists.txt
@@ -10,6 +10,8 @@ set(SRC_FILES
     src/NDArray.cpp
     src/ReleaseGlobalInterpreterLock.cpp
     src/UninstallTrace.cpp
+    src/PythonSink.cpp
+    src/PythonStdoutChannel.cpp
     src/WrapperHelpers.cpp
     src/Converters/CloneToNDArray.cpp
     src/Converters/DateAndTime.cpp
@@ -34,6 +36,8 @@ set(INC_FILES
     inc/MantidPythonInterface/core/NDArray.h
     inc/MantidPythonInterface/core/PropertyWithValueExporter.h
     inc/MantidPythonInterface/core/PythonObjectInstantiator.h
+    inc/MantidPythonInterface/core/PythonSink.h
+    inc/MantidPythonInterface/core/PythonStdoutChannel.h
     inc/MantidPythonInterface/core/ReleaseGlobalInterpreterLock.h
     inc/MantidPythonInterface/core/StlExportDefinitions.h
     inc/MantidPythonInterface/core/TypedValidatorExporter.h
