+++ Framework/PythonInterface/core/src/PythonStdoutChannel.cpp
@@ -4,23 +4,19 @@
 //   NScD Oak Ridge National Laboratory, European Spallation Source,
 //   Institut Laue - Langevin & CSNS, Institute of High Energy Physics, CAS
 // SPDX - License - Identifier: GPL - 3.0 +
-#include "MantidKernel/PythonSink.h"
+#include "MantidPythonInterface/core/PythonStdoutChannel.h"
+#include <Poco/Message.h>
+#include <iostream>
 
-#include <boost/format.hpp>
 #include <boost/iostreams/categories.hpp> // sink_tag
 #include <iosfwd>                         // streamsize
 
-#include <Python.h>
+#include <fstream>
 
 // test_ostream  std::cout
 
-std::streamsize pysys_stdout_sink::write(const char *s, std::streamsize n) {
-  // PySys_WriteStdout truncates to 1000 chars
-  static const std::streamsize MAXSIZE = 1000;
+namespace Poco {
 
-  std::streamsize written = std::min(n, MAXSIZE);
-  PySys_WriteStdout((boost::format("%%.%1%s") % written).str().c_str(), s);
-  // std::cout << s;
+PythonStdoutChannel::PythonStdoutChannel() : ConsoleChannel(pysys_stdout) {}
 
-  return written;
-}
+} // namespace Poco
